requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2019.3.9
chardet==3.0.4
decorator==4.3.0
defusedxml==0.6.0
Django==2.1.5
django-allauth==0.39.1
django-cors-headers==2.4.0
django-environ==0.4.5
django-simple-history==2.7.2
django-tenants==2.1.0
django-webpack-loader==0.6.0
djangorestframework==3.8.2
environ==1.0
gunicorn==19.9.0
idna==2.8
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.13.2
oauthlib==3.0.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==1.0.15
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
Pygments==2.3.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
simplegeneric==0.8.1
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.2
urllib3==1.24.2
wcwidth==0.1.7
